Replace isotropic Morlet monkey-patch with native filter basis
Move IsotropicMorletFilterBasis from the torch-harmonics monkey-patch
module into fme/core/disco/_filter_basis.py and register it directly in
get_filter_basis. Remove the now-unnecessary isotropic_filter.py and its
import from localnet.py.
